requirements.txt
boto3==1.34.34
apache-airflow==2.8.1
pyspark==3.5.0
pandas==2.0.3
dbt-redshift==1.7.4
python-dotenv==1.0.0
apache-airflow-providers-amazon==8.12.0
apache-airflow-providers-apache-spark==4.4.0
apache-airflow-providers-apache-hdfs==3.2.1 